// my-app/src/models/MissingComplaint.ts
export default interface IMissingComplaintFormData {
  complaintFiledBy: string
  complaintFiledAt: string
  caseClosedAt?: string
  username?: string
  name: string
  sex: string
  phoneNumber?: string
  age?: number
  height?: string
  weight?: string
  email?: string
  medications?: string
  allergies?: string
  conditions?: string
  birthmark?: string
  _id?: string
}